+++ test/helpers/fixture.js
@@ -153,7 +153,7 @@ export default class Fixture {
   // If you need to re-record cassettes, increment the date below and ensure it is one day in the future,
   // USPS only does "next-day" pickups including Saturday but not Sunday or Holidays.
   static basicPickup() {
-    const pickupDate = '2022-06-03';
+    const pickupDate = '2022-08-01';
 
     return {
       address: this.basicAddress(),
@@ -269,5 +269,47 @@ export default class Fixture {
     };
 
     return Buffer.from(JSON.stringify(data), 'utf8');
+  }
+
+  static carbonOffsetShipment() {
+    return {
+      to_address: {
+        name: 'Dr. Steve Brule',
+        street1: '179 N Harbor Dr',
+        city: 'Redondo Beach',
+        state: 'CA',
+        zip: '90277',
+        country: 'US',
+        phone: '8573875756',
+        email: 'dr_steve_brule@gmail.com',
+      },
+      from_address: {
+        name: 'EasyPost',
+        street1: '417 Montgomery Street',
+        street2: '5th Floor',
+        city: 'San Francisco',
+        state: 'CA',
+        zip: '94104',
+        country: 'US',
+        phone: '4153334445',
+        email: 'support@easypost.com',
+      },
+      parcel: {
+        length: 20.2,
+        width: 10.9,
+        height: 5,
+        weight: 65.9,
+      },
+    };
+  }
+
+  static oneCallBuyCarbonOffset() {
+    const shipment = this.carbonOffsetShipment();
+
+    shipment.service = 'Priority';
+    shipment.carrier_accounts = [this.uspsCarrierAccountId()];
+    shipment.carrier = this.usps();
+
+    return shipment;
   }
 }